+++ src/Kernel.php
@@ -10,6 +10,7 @@ use AvtoDev\JsonRpc\Errors\ErrorInterface;
 use AvtoDev\JsonRpc\Router\RouterInterface;
 use AvtoDev\JsonRpc\Responses\ErrorResponse;
 use AvtoDev\JsonRpc\Responses\ResponsesStack;
+use Illuminate\Contracts\Container\Container;
 use AvtoDev\JsonRpc\Requests\RequestInterface;
 use AvtoDev\JsonRpc\Responses\SuccessResponse;
 use AvtoDev\JsonRpc\Errors\MethodNotFoundError;
@@ -19,10 +20,17 @@ use AvtoDev\JsonRpc\Requests\ErroredRequestInterface;
 use AvtoDev\JsonRpc\Responses\ResponsesStackInterface;
 use AvtoDev\JsonRpc\Events\ErroredRequestDetectedEvent;
 use AvtoDev\JsonRpc\Events\RequestHandledExceptionEvent;
+use AvtoDev\JsonRpc\Requests\RequestInterface as RPCRequest;
+use AvtoDev\JsonRpc\MethodParameters\MethodParametersInterface;
 use Illuminate\Contracts\Events\Dispatcher as EventsDispatcher;
 
 class Kernel implements KernelInterface
 {
+    /**
+     * @var Container
+     */
+    protected $container;
+
     /**
      * @var RouterInterface
      */
@@ -36,13 +44,17 @@ class Kernel implements KernelInterface
     /**
      * Kernel constructor.
      *
+     * @param Container        $container
      * @param RouterInterface  $router
      * @param EventsDispatcher $events
      */
-    public function __construct(RouterInterface $router, EventsDispatcher $events)
+    public function __construct(Container $container, RouterInterface $router, EventsDispatcher $events)
     {
-        $this->router = $router;
-        $this->events = $events;
+        $this->container = $container;
+        $this->router    = $router;
+        $this->events    = $events;
+
+        $this->bindParametersParser($container);
     }
 
     /**
@@ -54,39 +66,115 @@ class Kernel implements KernelInterface
 
         foreach ($requests->all() as $request) {
             if ($request instanceof ErroredRequestInterface) {
-                $this->events->dispatch(new ErroredRequestDetectedEvent($request));
-
-                $responses->push(new ErrorResponse($request->getId(), $request->getError()));
+                // Process any requests that have errors
+                $this->processErroredRequest($request, $responses);
             } elseif ($request instanceof RequestInterface) {
                 if ($this->router->methodExists($request->getMethod())) {
                     try {
-                        $result = $this->router->handle($request);
+                        // Bind request instance into container
+                        $this->bindRpcRequest($request);
 
-                        if ($request->isNotification() === false) {
-                            $responses->push(new SuccessResponse($request->getId(), $result));
-                        }
-
-                        $this->events->dispatch(new RequestHandledEvent($request));
+                        // Handle current request and push it in Responses Stack
+                        $this->handleRequest($request, $responses);
                     } catch (Throwable $e) {
-                        $this->events->dispatch(new RequestHandledExceptionEvent($request, $e));
-
-                        $responses->push(
-                            new ErrorResponse(
-                                $request->getId(),
-                                $e instanceof ErrorInterface
-                                    ? $e
-                                    : new InternalError(null, (int) $e->getCode(), $e, $e)
-                            )
-                        );
+                        // Catch any error that occurred while executing the request
+                        $this->catchErrorOnHandle($request, $responses, $e);
                     }
                 } else {
-                    if ($request->isNotification() === false) {
-                        $responses->push(new ErrorResponse($request->getId(), new MethodNotFoundError));
-                    }
+                    $this->pushErrorIfNotNotification($request, $responses);
                 }
             }
         }
 
         return $responses;
+    }
+
+    /**
+     * @param Container $container
+     */
+    protected function bindParametersParser(Container $container): void
+    {
+        $container->resolving(
+            MethodParametersInterface::class,
+            function (MethodParametersInterface $parameters) use ($container): void {
+                $request = $container->make(RPCRequest::class);
+
+                if ($request instanceof RPCRequest) {// @todo: Make test for this condition
+                    $parameters->parse($request->getParams());
+                }
+            }
+        );
+    }
+
+    /**
+     * @param ErroredRequestInterface $request
+     * @param ResponsesStackInterface $responses
+     */
+    protected function processErroredRequest(ErroredRequestInterface $request, ResponsesStackInterface $responses): void
+    {
+        $this->events->dispatch(new ErroredRequestDetectedEvent($request));
+
+        $responses->push(new ErrorResponse($request->getId(), $request->getError()));
+    }
+
+    /**
+     * @param RPCRequest              $request
+     * @param ResponsesStackInterface $responses
+     *
+     * @throws Throwable
+     */
+    protected function handleRequest(RequestInterface $request, ResponsesStackInterface $responses): void
+    {
+        $result = $this->router->call($request->getMethod());
+
+        if ($request->isNotification() === false) {
+            $responses->push(new SuccessResponse($request->getId(), $result));
+        }
+
+        $this->events->dispatch(new RequestHandledEvent($request));
+    }
+
+    /**
+     * @param RPCRequest              $request
+     * @param ResponsesStackInterface $responses
+     * @param Throwable               $e
+     */
+    protected function catchErrorOnHandle(RequestInterface $request,
+                                          ResponsesStackInterface $responses,
+                                          Throwable $e): void
+    {
+        $this->events->dispatch(new RequestHandledExceptionEvent($request, $e));
+
+        if ($request->isNotification() === false) {
+            $responses->push(
+                new ErrorResponse(
+                    $request->getId(),
+                    $e instanceof ErrorInterface
+                        ? $e
+                        : new InternalError(null, (int) $e->getCode(), $e, $e)
+                )
+            );
+        }
+    }
+
+    /**
+     * @param RPCRequest              $request
+     * @param ResponsesStackInterface $responses
+     */
+    protected function pushErrorIfNotNotification(RequestInterface $request, ResponsesStackInterface $responses): void
+    {
+        if ($request->isNotification() === false) {
+            $responses->push(new ErrorResponse($request->getId(), new MethodNotFoundError));
+        }
+    }
+
+    /**
+     * @param RPCRequest $request
+     */
+    protected function bindRpcRequest(RequestInterface $request): void
+    {
+        $this->container->bind(RPCRequest::class, static function () use ($request): RPCRequest {
+            return $request;
+        });
     }
 }