+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## Unreleased
+
+### Changed
+
+- Renamed `handle` method to `call` in Router
+- Refactoring method `Kernel::handle`
+- Some tests in connection with innovations
+
 ## v1.0.0
 
 ### Added